client/X11: Reintroduce handling for relieving filename restriction
Commit 25e2dd7ab0acf351a016fdd5f08b7b4318f2a6c3 introduced a way to
relieve the filename restriction, which ensured that file lists
containing files with names, that have characters, which are not allowed
on MS Windows.
This relief handling kicked in, when xfreerdp did not connect to MS
Windows RDS.
Commit f3f7c7dfe0b20e6dbe306a397e28ed7abca65f81 got rid of any
wClipboardDelegate occurrence in xf_cliprdr.c, which also got rid of the
filename restriction relieve handling.
However, the relief handling was not added back, so do it now.