new [libfreerdp]: use compile time paths in FreeRDP cmake module
The exported paths for

* FreeRDP_PLUGIN_DIR
* FreeRDP_PROXY_PLUGIN_DIR
* FreeRDP_EXTENSION_DIR

introduced in 7959215d52d21b4d6c1cef737436aa63535665c8 were relative to
the installation of the cmake module because of the use of CMakePackageConfigHelpers.
This could lead to different paths for configuration and runtime -
causing channels not to be found and loaded.

Instead of using paths relative to the cmake module set the paths
configured during configure/compile for plugins and extension.

This way the exported paths match the paths in build-config.h and are the
same used in the pkg-config variables.